+++ src/ifba/ads/controll/H2CRUD.java
@@ -5,8 +5,8 @@ import ifba.ads.model.UnidadeMovel;
 public interface H2CRUD {
 	
 	abstract void inserir(UnidadeMovel unidade);
-	abstract void atualizar();
-	abstract void deletar();
+	abstract void atualizar(UnidadeMovel unidade);
+	abstract void deletar(UnidadeMovel unidade);
 	abstract void consultar();
 	
 }